support pytorch autolog even w/o lightning installed
