+++ mlflow/pytorch/__init__.py
@@ -16,6 +16,7 @@ import warnings
 import numpy as np
 import pandas as pd
 from packaging.version import Version
+from functools import partial
 import posixpath
 
 import mlflow
@@ -1041,99 +1042,26 @@ def autolog(
         PyTorch autologged MLflow entities
     """
     import atexit
-    import time
-    import pytorch_lightning as pl
-    from mlflow.tracking import MlflowClient
-    from mlflow.pytorch._pytorch_autolog import patched_fit, IN_FIT
-    from mlflow.entities import Param, Metric
-    from mlflow.tensorflow import (
+    from mlflow.pytorch._lightning_autolog import patched_fit, HAVE_LIGHTNING
+    from mlflow.pytorch._pytorch_autolog import (
+        patched_add_event,
+        patched_add_hparams,
+        patched_add_summary,
         _flush_queue,
-        _metric_queue,
-        _metric_queue_lock,
-        _thread_pool,
-        _MAX_METRIC_QUEUE_SIZE,
     )
 
-    def _add_to_queue(key, value, step, time, run_id):
-        """
-        Add a metric to the metric queue. Flush the queue if it exceeds the
-        max queue size.
-        """
-        met = Metric(key=key, value=value, timestamp=time, step=step)
-        with _metric_queue_lock:
-            _metric_queue.append((run_id, met))
-            if len(_metric_queue) > _MAX_METRIC_QUEUE_SIZE:
-                _thread_pool.submit(_flush_queue)
+    if HAVE_LIGHTNING:
+        import pytorch_lightning as pl
 
-    # lightening
-    safe_patch(FLAVOR_NAME, pl.Trainer, "fit", patched_fit, manage_run=True)
-
-    FAILED = object()
-    autolog_run = None
-
-    def _get_run():
-        nonlocal autolog_run
-        if autolog_run is None:
-            autolog_run = mlflow.active_run()
-        if autolog_run is None:
-            try:
-                autolog_run = mlflow.start_run()
-            except Exception:
-                # don't try again.
-                autolog_run = FAILED
-        if autolog_run not in (None, FAILED):
-            return autolog_run
-        return None
-
-    def patched_add_hparams(original, self, hparam_dict, metric_dict, *args, **kwargs):
-        """use a synchronous call here since this is going to get called very infrequently."""
-
-        run = _get_run()
-
-        if not IN_FIT and run is not None and hparam_dict:
-            run_id = run.info.run_id
-            # str() is required by mlflow :(
-            params_arr = [Param(key, str(value)) for key, value in hparam_dict.items()]
-            metrics_arr = [
-                Metric(key, value, int(time.time() * 1000), 0) for key, value in metric_dict.items()
-            ]
-            MlflowClient().log_batch(run_id=run_id, metrics=metrics_arr, params=params_arr, tags=[])
-
-        original(self, hparam_dict, metric_dict, *args, **kwargs)
-
-    def patched_add_event(original, self, event, *args, **kwargs):
-        run = _get_run()
-        if (
-            not IN_FIT
-            and run is not None
-            and event.WhichOneof("what") == "summary"
-            and log_every_n_step
-        ):
-            summary = event.summary
-            global_step = args[0] if len(args) > 0 else kwargs.get("global_step", None)
-            global_step = global_step or 0
-            for v in summary.value:
-                if v.HasField("simple_value"):
-                    if global_step % log_every_n_step == 0:
-                        _add_to_queue(
-                            key=v.tag,
-                            value=v.simple_value,
-                            step=global_step,
-                            time=int((event.wall_time or time.time()) * 1000),
-                            run_id=run.info.run_id,
-                        )
-
-        return original(self, event, *args, **kwargs)
-
-    def patched_add_summary(original, self, *args, **kwargs):
-        result = original(self, *args, **kwargs)
-        _flush_queue()
-        return result
+        safe_patch(FLAVOR_NAME, pl.Trainer, "fit", patched_fit, manage_run=True)
 
     import torch.utils.tensorboard.writer
 
     safe_patch(
-        FLAVOR_NAME, torch.utils.tensorboard.writer.FileWriter, "add_event", patched_add_event
+        FLAVOR_NAME,
+        torch.utils.tensorboard.writer.FileWriter,
+        "add_event",
+        partial(patched_add_event, mlflow_log_every_n_step=log_every_n_step),
     )
     safe_patch(
         FLAVOR_NAME, torch.utils.tensorboard.writer.FileWriter, "add_summary", patched_add_summary
